+++ src/main/resources/application.yml
@@ -31,7 +31,14 @@ spring:
     check-template-location: true
     cache: false
     mvc:
-      static-path-pattern: /static/**
+      static-path-pattern: "/static/**"
+
+  web:
+    resources:
+      static-locations: classpath:/static/
+  mvc:
+    contentnegotiation:
+      favor-path-extension: true
 
 jwt:
   secret-key: ${JWT_SECRET_KEY}
