+++ src/main/resources/templates/user/ChatList.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Katsee</title>
-    <link rel="stylesheet" href="/static/user/ChatList.css">
+    <link rel="stylesheet" href="/user/ChatList.css">
 </head>
 <body>
     <h2 class="header">채팅</h2>
