+++ src/main/resources/templates/user/FriendRecommend.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Katsee</title>
-    <link rel="stylesheet" href="/static/user/FriendRecommend.css">
+    <link rel="stylesheet" href="/user/FriendRecommend.css">
 </head>
 <body>
     <h2 class="header">친구 추천</h2>
