+++ src/main/resources/templates/user/FriendRequestList.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Katsee</title>
-    <link rel="stylesheet" href="/static/user/FriendRequestList.css">
+    <link rel="stylesheet" href="/user/FriendRequestList.css">
 </head>
 <body>
     <h2 class="header">친구 요청 목록</h2>
