+++ src/main/resources/templates/user/JoinSecond.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Katsee</title>
-    <link rel="stylesheet" href="/static/user/JoinSecond.css">
+    <link rel="stylesheet" href="/user/JoinSecond.css">
 </head>
 <body>
     <div class="container">
