+++ src/main/resources/templates/user/MemberEighthDetails.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Katsee</title>
-    <link rel="stylesheet" href="/static/user/MemberEighthDetails.css">
+    <link rel="stylesheet" href="/user/MemberEighthDetails.css">
 </head>
 <body>
     <div class="container">
