+++ src/main/resources/templates/user/MemberThirdDetails.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Katsee</title>
-    <link rel="stylesheet" href="/static/user/MemberThirdDetails.css">
+    <link rel="stylesheet" href="/user/MemberThirdDetails.css">
 </head>
 <body>
     <div class="container">
