+++ src/main/resources/templates/user/UserHome.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Katsee</title>
-    <link rel="stylesheet" href="/static/user/UserHome.css">
+    <link rel="stylesheet" href="/user/UserHome.css">
 </head>
 <body>
     <div class="container">
