+++ src/main/resources/templates/user/UserNotice.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Katsee</title>
-    <link rel="stylesheet" href="/static/user/UserNotice.css">
+    <link rel="stylesheet" href="/user/UserNotice.css">
 </head>
 <body>
     <div class="container">
